+++ tests/testapp/src/app/mod.rs
@@ -10,7 +10,11 @@ use testapp::{AppReadRequest, AppState, AppWriteRequest};
 
 mod snapshot_io;
 
-pub async fn new(log: impl RaftLogStore, ballot: impl RaftBallotStore, driver: lolraft::RaftDriver) -> Result<RaftProcess> {
+pub async fn new(
+    log: impl RaftLogStore,
+    ballot: impl RaftBallotStore,
+    driver: lolraft::RaftDriver,
+) -> Result<RaftProcess> {
     let app_main = AppMain::new();
     let process = RaftProcess::new(app_main, log, ballot, driver).await?;
     Ok(process)
